🔥(xblock) remove extra python requirements file for oee
This fork makes sure Open edX does not fail if the configurable LTI
consumer XBlock is not installed. So the responsibility to install
should be in openedx-docker.

More importantly, it will also save us from having to make a new
release of this fork each time we change the version of the configurable
XBlock!